Moved file structure around in components and updated imports still needs testing
