+++ src/app/about/page.js
@@ -1,4 +1,4 @@
-import TypewriterMain from "@/components/TypewriterMain"
+import TypewriterMain from "@/components/About/TypewriterMain"
 
 export default function Page() {
     return (
