+++ src/app/blog/[slug]/page.js
@@ -1,5 +1,5 @@
-import BlogPostFetcher from "@/components/BlogPostFetcher";
-import ClientBlogLayout from "@/components/ClientBlogLayout";
+import BlogPostFetcher from "@/components/Blog/BlogPostFetcher";
+import ClientBlogLayout from "@/components/Blog/ClientBlogLayout";
 
 export default async function Page({ params }) {
     const { slug } = await params; // Next yells at you if you don't destructure this from params
