+++ src/app/page.js
@@ -1,4 +1,4 @@
-import Main from "@/components/Main";
+import Main from "@/components/Landing/Main";
 
 export default function Home() {
   return (
